Reboot the VM if we missed the boot menu.
Will-fix: #10777

+++ features/step_definitions/common_steps.rb
@@ -307,17 +307,29 @@ Given /^Tails is at the boot menu( after rebooting)?$/ do |reboot|
     end
   end
   add_after_scenario_hook { kill_tab_spammer.call }
-  # Our UEFI bootloader has the interesting "feature" that pressing
-  # any button will open its setup menu, so we have to exit the setup,
-  # and to not have the TAB spammer potentially interfering we pause
-  # it meanwhile.
-  if @os_loader == 'UEFI'
-    @screen.wait('UEFIBootLoaderSetup.png', boot_timeout)
-    Process.kill("TSTP", tab_spammer.pid)
-    @screen.type(Sikuli::Key.ENTER)
-    Process.kill("CONT", tab_spammer.pid)
+  dealt_with_uefi_setup = false
+  try_for(boot_timeout) do
+    begin
+      # Our UEFI bootloader has the interesting "feature" that
+      # pressing any button will open its setup menu, so we have to
+      # exit the setup, and to not have the TAB spammer potentially
+      # interfering we pause it meanwhile.
+      if not(dealt_with_uefi_setup) && @os_loader == 'UEFI'
+        @screen.find('UEFIBootLoaderSetup.png')
+        Process.kill("TSTP", tab_spammer.pid)
+        @screen.type(Sikuli::Key.ENTER)
+        Process.kill("CONT", tab_spammer.pid)
+        dealt_with_uefi_setup = true
+      end
+      @screen.find(boot_menu_cmdline_image)
+    rescue FindFailed => e
+      if @screen.exist('TailsGreeter.png')
+        $vm.reset
+        dealt_with_uefi_setup = false
+      end
+      raise e
+    end
   end
-  @screen.wait(boot_menu_cmdline_image, boot_timeout)
   kill_tab_spammer.call
   # Ensure that we're back at the boot splash
   @screen.type(Sikuli::Key.ESC)
